# api/docs/swagger.yaml
definitions:
  models.Capacity:
    properties:
      cpu:
        type: number
      memory:
        type: integer
    type: object
...
  models.Quota:
    properties:
      capacity:
        $ref: '#/definitions/models.Capacity'
      group_id:
        type: string
      id:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Go-server for PAC backend apis in gin
  title: Pac-go-server APIs
paths:
...
  /api/v1/groups/{id}/quota:
    delete:
      consumes:
      - application/json
      description: Delete quota
      parameters:
      - description: group-id where quota has to be deleted
        in: path
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete quota
      tags:
      - quota
    get:
      consumes:
      - application/json
      description: Get quota
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: quota-id to be fetched
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get quota
      tags:
      - quota
    post:
      consumes:
      - application/json
      description: Create quota
      parameters:
      - description: Create quota
        in: body
        name: quota
        required: true
        schema:
          $ref: '#/definitions/models.Quota'
      - description: group-id where quota has to be created
        in: path
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create quota
      tags:
      - quota
    put:
      consumes:
      - application/json
      description: Update quota
      parameters:
      - description: Update quota
        in: body
        name: quota
        required: true
        schema:
          $ref: '#/definitions/models.Quota'
      - description: group-id where quota has to be updated
        in: path
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update quota
      tags:
      - quota
...
  /api/v1/quota:
    get:
      consumes:
      - application/json
      description: Get user quota
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get user quota
      tags:
      - quota
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all users
      tags:
      - user
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get user based on given id
      parameters:
      - description: user-id for user to be fetched
        in: path
        name: id
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get user
      tags:
      - user
swagger: "2.0"